site tab refactor (#1173)
Initial refactoring for progress on #1052

Moving each class to separate files, and switching from Vector to List.

Changes to sorting and Filtering will be in separate PR(s)

(cherry picked from commit c79d1dea7a8fb27575083e2a849ad083faef73e1)
